refactor(mb-api): consolidate indicateurs routes per entity
PR review (#1, #9, #11, #12, #13, #7):
- Merge getIndicateurs.ts + getIndicateurById.ts into indicateur/routes.ts
- Compose IndicateurListApiModelSchema locally in route file (factory +
  entity from mb-shared) instead of importing the pre-composed schema
- Validate cursor with Zod regex (^\d+$): non-numeric cursors return 400
- Translate French variable q -> searchTerm
- Keep route-imported-first ordering in tests (Zod 4 prototype extension
  does not propagate to ZodObject instances created before the extend
  call, so mb-shared schemas must be evaluated after @hono/zod-openapi
  loads). Comment updated to reflect the real reason.
