+++ apps/mb-api/src/app.ts
@@ -4,8 +4,7 @@ import { cors } from 'hono/cors'
 
 import { env } from '@/env'
 import { health } from '@/healthcheck/routes/health'
-import { getIndicateurById } from '@/indicateur/routes/getIndicateurById'
-import { getIndicateurs } from '@/indicateur/routes/getIndicateurs'
+import { indicateurRoutes } from '@/indicateur/routes'
 import { sharedMessage } from '@/shared/routes/sharedMessage'
 
 export const app = new OpenAPIHono()
@@ -15,8 +14,7 @@ app.use('*', cors({ origin: env.CORS_ORIGINS, credentials: true }))
 app.get('/', (context) => context.json({ hello: 'world' }))
 app.route('/', health)
 app.route('/', sharedMessage)
-app.route('/', getIndicateurs)
-app.route('/', getIndicateurById)
+app.route('/', indicateurRoutes)
 
 app.doc('/openapi.json', {
   openapi: '3.0.0',
