+++ apps/mb-api/src/indicateur/routes.test.ts
@@ -1,12 +1,22 @@
-// Import the route first so @hono/zod-openapi extends the Zod prototype
-// before mb-shared schemas are evaluated.
-import { getIndicateurs } from '@/indicateur/routes/getIndicateurs'
-import { indicateurListApiModelSchema } from '@pilote/mb-shared'
+// Route module imported first so @hono/zod-openapi extends Zod's prototype
+// before any mb-shared schemas are evaluated. Once mb-shared switches to
+// schema factories (no module-level z.object() calls), this can be relaxed.
+import {
+  ErrorApiModelSchema,
+  INDICATEUR_NOT_FOUND_CODE,
+  indicateurRoutes,
+} from '@/indicateur/routes'
+import {
+  createPaginatedApiListSchema,
+  indicateurApiModelSchema,
+} from '@pilote/mb-shared'
 import { describe, expect, it } from 'vitest'
+
+const indicateurListApiModelSchema = createPaginatedApiListSchema(indicateurApiModelSchema)
 
 describe('GET /indicateurs', () => {
   it('retourne la première page sans filtre (5 items, hasMore=true)', async () => {
-    const response = await getIndicateurs.request('/indicateurs')
+    const response = await indicateurRoutes.request('/indicateurs')
     const body = indicateurListApiModelSchema.parse(await response.json())
 
     expect(response.status).toBe(200)
@@ -18,25 +28,22 @@ describe('GET /indicateurs', () => {
   })
 
   it('paginate end-to-end : suit le cursor jusqu\'à hasMore=false', async () => {
-    const r1 = await getIndicateurs.request('/indicateurs')
+    const r1 = await indicateurRoutes.request('/indicateurs')
     const b1 = indicateurListApiModelSchema.parse(await r1.json())
     expect(b1.pagination.hasMore).toBe(true)
-    expect(b1.pagination.cursor).not.toBeNull()
 
-    const r2 = await getIndicateurs.request(
+    const r2 = await indicateurRoutes.request(
       `/indicateurs?cursor=${b1.pagination.cursor}`,
     )
     const b2 = indicateurListApiModelSchema.parse(await r2.json())
-
     expect(b2.items).toHaveLength(3)
-    expect(b2.items.every((i) => i.id > 5)).toBe(true)
     expect(b2.pagination.hasMore).toBe(false)
     expect(b2.pagination.cursor).toBeNull()
     expect(b2.total).toBe(8)
   })
 
   it('cursor au-delà du dernier id retourne une page vide', async () => {
-    const response = await getIndicateurs.request('/indicateurs?cursor=9999')
+    const response = await indicateurRoutes.request('/indicateurs?cursor=9999')
     const body = indicateurListApiModelSchema.parse(await response.json())
 
     expect(response.status).toBe(200)
@@ -45,8 +52,13 @@ describe('GET /indicateurs', () => {
     expect(body.pagination.cursor).toBeNull()
   })
 
+  it('cursor invalide (non numérique) retourne 400', async () => {
+    const response = await indicateurRoutes.request('/indicateurs?cursor=abc')
+    expect(response.status).toBe(400)
+  })
+
   it('filtre par statut', async () => {
-    const response = await getIndicateurs.request('/indicateurs?statut=archive')
+    const response = await indicateurRoutes.request('/indicateurs?statut=archive')
     const body = indicateurListApiModelSchema.parse(await response.json())
 
     expect(response.status).toBe(200)
@@ -56,7 +68,7 @@ describe('GET /indicateurs', () => {
   })
 
   it('filtre par recherche (nom, case-insensitive)', async () => {
-    const response = await getIndicateurs.request('/indicateurs?recherche=fibre')
+    const response = await indicateurRoutes.request('/indicateurs?recherche=fibre')
     const body = indicateurListApiModelSchema.parse(await response.json())
 
     expect(response.status).toBe(200)
@@ -65,7 +77,7 @@ describe('GET /indicateurs', () => {
   })
 
   it('retourne une liste vide quand aucune correspondance', async () => {
-    const response = await getIndicateurs.request(
+    const response = await indicateurRoutes.request(
       '/indicateurs?recherche=zzzzznoresult',
     )
     const body = indicateurListApiModelSchema.parse(await response.json())
@@ -75,5 +87,30 @@ describe('GET /indicateurs', () => {
     expect(body.total).toBe(0)
     expect(body.pagination.hasMore).toBe(false)
     expect(body.pagination.cursor).toBeNull()
+  })
+})
+
+describe('GET /indicateurs/:id', () => {
+  it('retourne un indicateur par id', async () => {
+    const response = await indicateurRoutes.request('/indicateurs/1')
+    const body = indicateurApiModelSchema.parse(await response.json())
+
+    expect(response.status).toBe(200)
+    expect(body.id).toBe(1)
+    expect(body.nom).toBe('Taux de chômage')
+  })
+
+  it('retourne 404 avec un code stable si l\'indicateur n\'existe pas', async () => {
+    const response = await indicateurRoutes.request('/indicateurs/9999')
+    const body = ErrorApiModelSchema.parse(await response.json())
+
+    expect(response.status).toBe(404)
+    expect(body.code).toBe(INDICATEUR_NOT_FOUND_CODE)
+    expect(body.message).toBe('Indicateur introuvable')
+  })
+
+  it('retourne 400 si l\'id n\'est pas un nombre', async () => {
+    const response = await indicateurRoutes.request('/indicateurs/abc')
+    expect(response.status).toBe(400)
   })
 })
